+++ examples/py_systemd/code/sock.py
@@ -1,0 +1,13 @@
+import os
+import sys
+import socket
+
+assert os.environ["LISTEN_FDS"] == "1"
+assert os.environ["LISTEN_NAMES"] == "input_port"
+
+sock = socket.socket(fileno=3)
+while True:
+    s, a = sock.accept()
+    s.send(b'hello')
+    s.close()
+